self_learning/projetc1_bot/requirements.txt
attrs==21.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
cryptography==35.0.0
idna==3.2
pkg_resources==0.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==21.0.0
python-decouple==3.4
requests==2.26.0
service-identity==21.1.0
six==1.16.0
TelegramBotAPI==0.3.2
urllib3==1.26.7
